fix: Handle WebSocket connection errors and retry
This commit modifies the `NetworkingService` class to handle WebSocket connection errors. When a connection error occurs, the error message is logged to the console. If the WebSocket is not active and the retry counter is below 50, the `connect()` method is called to retry the connection. This ensures that the WebSocket connection is reestablished in case of errors. The previous code that added an error message to the UI has been commented out.